move pilToNSImage to drawBot.misc
